Code: Update deletePe logic.
We delete the page element by finding its deleteIndex in the data list.

After that, if there remains no page element, we simply finish.
If there are remaining page element(s), we try to select the one that
is currently at deleteIndex, if that it not possible(we deleted the
last element), then we select the one that is immediately before it.